# Small performance improvements for AddNamespaceDeclaration