--- a/test/DocumentFormat.OpenXml.Benchmarks/OpenXmlElementAddNamespaceDeclarationTests.cs
+++ b/test/DocumentFormat.OpenXml.Benchmarks/OpenXmlElementAddNamespaceDeclarationTests.cs
@@ -1,0 +1,31 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using BenchmarkDotNet.Attributes;
+using DocumentFormat.OpenXml.Drawing.ChartDrawing;
+
+namespace DocumentFormat.OpenXml.Benchmarks
+{
+    public class OpenXmlElementAddNamespaceDeclarationTests
+    {
+        [IterationSetup]
+        public void Setup()
+        {
+            _element = new TextBody();
+        }
+
+        [Benchmark]
+        public void AddNamespaceDeclaration()
+        {
+            _element.AddNamespaceDeclaration("test","test");
+        }
+
+        [Benchmark]
+        public void OldAddNamespaceDeclaration()
+        {
+            _element.OldAddNamespaceDeclaration("test", "test");
+        }
+
+        private OpenXmlElement _element;
+    }
+}