|                     Method |     Mean |     Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------- |---------:|----------:|---------:|------:|------:|------:|----------:|
|    AddNamespaceDeclaration | 3.665 us | 0.5911 us | 1.658 us |     - |     - |     - |         - |
| OldAddNamespaceDeclaration | 4.303 us | 0.8884 us | 2.462 us |     - |     - |     - |         - |


Because the method execution time is too short, it is difficult to observe the intensity of optimization.


Benchmark code: https://github.com/dotnet-campus/Open-XML-SDK/blob/5d5c307a55066fccb3913d28bf945b51e0a6fcde/test/DocumentFormat.OpenXml.Benchmarks/OpenXmlElementAddNamespaceDeclarationTests.cs