@lindexi 2020-08-07
@twsouthwick  I added the benchmark code.